requirements.txt
python-dotenv
spotipy
ytmusicapi
tqdm